Move literature lookup table
